+++ coloraide/colors/_space.py
@@ -173,7 +173,8 @@ class Space(contrast.Contrast, interpolate.Interpolate, distance.Distance, gamut
         a = util.no_nan(self.alpha)
         alpha = alpha is not False and (alpha is True or a < 1.0)
 
-        coords = util.no_nan(self.fit_coords() if fit else self.coords())
+        method = None if not isinstance(fit, str) else fit
+        coords = util.no_nan(self.fit_coords(method=method) if fit else self.coords())
         template = "color({} {} {} {} / {})" if alpha else "color({} {} {} {})"
         values = [
             util.fmt_float(coords[0], precision),