+++ coloraide/colors/hwb.py
@@ -65,11 +65,6 @@ class HWB(Cylindrical, Space):
         if isinstance(color, Space):
             self.hue, self.whiteness, self.blackness = color.convert(self.space()).coords()
             self.alpha = color.alpha
-        elif isinstance(color, str):
-            values = self.match(color)[0]
-            if values is None:
-                raise ValueError("'{}' does not appear to be a valid color".format(color))
-            self.hue, self.whiteness, self.blackness, self.alpha = values
         elif isinstance(color, (list, tuple)):
             if not (3 <= len(color) <= 4):
                 raise ValueError("A list of channel values should be of length 3 or 4.")