+++ coloraide/colors/lab.py
@@ -77,11 +77,6 @@ class LAB(Space):
         if isinstance(color, Space):
             self.lightness, self.a, self.b = color.convert(self.space()).coords()
             self.alpha = color.alpha
-        elif isinstance(color, str):
-            values = self.match(color)[0]
-            if values is None:
-                raise ValueError("'{}' does not appear to be a valid color".format(color))
-            self.lightness, self.a, self.b, self.alpha = values
         elif isinstance(color, (list, tuple)):
             if not (3 <= len(color) <= 4):
                 raise ValueError("A list of channel values should be of length 3 or 4.")