+++ coloraide/colors/srgb.py
@@ -101,11 +101,6 @@ class SRGB(Space):
         if isinstance(color, Space):
             self.red, self.green, self.blue = color.convert(self.space()).coords()
             self.alpha = color.alpha
-        elif isinstance(color, str):
-            values = self.match(color)[0]
-            if values is None:
-                raise ValueError("'{}' does not appear to be a valid color".format(color))
-            self.red, self.green, self.blue, self.alpha = values
         elif isinstance(color, (list, tuple)):
             if not (3 <= len(color) <= 4):
                 raise ValueError("A list of channel values should be of length 3 or 4.")