+++ coloraide/css/colors/lab.py
@@ -44,7 +44,8 @@ class LAB(generic.LAB):
 
         a = util.no_nan(self.alpha)
         alpha = alpha is not False and (alpha is True or a < 1.0)
-        coords = util.no_nan(self.fit_coords() if fit else self.coords())
+        method = None if not isinstance(fit, str) else fit
+        coords = util.no_nan(self.fit_coords(method=method) if fit else self.coords())
 
         if alpha:
             template = "lab({}%, {}, {}, {})" if options.get("comma") else "lab({}% {} {} / {})"